refactor: move Shared Drive params and Office extraction to base GoogleDriveTools
Add corpora, supports_all_drives, include_items_from_all_drives, and
drive_id params to GoogleDriveTools. Defaults unchanged for backward
compatibility. Context provider now uses base toolkit with explicit
params instead of a forked subclass.

- Add 4 new params to GoogleDriveTools.__init__() for Shared Drive support
- Move Office extraction helpers (_extract_docx_text, etc.) to base toolkit
- Delete context/gdrive/tools.py (AllDrivesGoogleDriveTools fork)
- Update GoogleDriveContextProvider to use GoogleDriveTools(corpora="allDrives")
- Update cookbook docstring to reflect new architecture
